backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0.0
pydantic
Pillow
ExifRead
numpy
alembic
typer
pytest
pytest-cov
imagehash
sentence-transformers
open_clip_torch
torch
torchvision
faiss-cpu
